#4 container-presenter pattern
