+++ src/screens/RankingScreen/index.js
@@ -1,0 +1,1 @@
+export {RankingScreenContainer as RankingScreen} from './RankingScreenContainer';
