+++ src/screens/RankingScreen/ranking_dummy.js
